refactor: 폴더명을 복수형으로 변경
