--- src/component/@common/GridLayout/GridLayout.stories.jsx
@@ -1,41 +1,0 @@
-import GridLayout from 'component/@common/GridLayout/GridLayout';
-import { Product } from 'component';
-
-export default {
-  title: 'Component/@Common/GridLayout',
-  component: GridLayout,
-  parameters: {
-    layout: 'centered',
-  },
-};
-
-const Template = args => <GridLayout {...args}></GridLayout>;
-
-const DefaulGridLayout = Template.bind({});
-
-DefaulGridLayout.args = {
-  children: [
-    {
-      image: 'https://i.pinimg.com/474x/4a/47/d5/4a47d5956eb090ff702e3b2cb47fdf98.jpg',
-      name: '사과',
-      price: 1300,
-    },
-    {
-      image: 'https://i.pinimg.com/474x/4a/47/d5/4a47d5956eb090ff702e3b2cb47fdf98.jpg',
-      name: '사과',
-      price: 1300,
-    },
-    {
-      image: 'https://i.pinimg.com/474x/4a/47/d5/4a47d5956eb090ff702e3b2cb47fdf98.jpg',
-      name: '사과',
-      price: 1300,
-    },
-    {
-      image: 'https://i.pinimg.com/474x/4a/47/d5/4a47d5956eb090ff702e3b2cb47fdf98.jpg',
-      name: '사과',
-      price: 1300,
-    },
-  ].map(product => <Product key={product.id} {...product} />),
-};
-
-export { DefaulGridLayout };
