--- src/component/@common/index.js
@@ -1,4 +1,0 @@
-export { default as Button } from 'component/@common/Button/Button';
-export { default as GridLayout } from 'component/@common/GridLayout/GridLayout';
-export { default as Header } from 'component/@common/Header/Header';
-export { default as Spinner } from 'component/@common/Spinner/Spinner';
