--- src/component/LoadingSpinner/LoadingSpinner.stories.jsx
@@ -1,15 +1,0 @@
-import { LoadingSpinner } from 'component';
-
-export default {
-  title: 'Component/LoadingSpinner',
-  component: LoadingSpinner,
-  parameters: {
-    layout: 'centered',
-  },
-};
-
-const Template = args => <LoadingSpinner {...args} />;
-
-const DefaultLoadingSpinner = Template.bind({});
-
-export { DefaultLoadingSpinner };
