--- src/component/Result/Result.stories.jsx
@@ -1,21 +1,0 @@
-import { Result } from 'component';
-
-export default {
-  title: 'Component/Result',
-  component: Result,
-  parameters: {
-    layout: 'centered',
-  },
-};
-
-const Template = args => <Result {...args} />;
-
-const ShoppingCartResult = Template.bind({});
-
-ShoppingCartResult.args = {
-  title: '결제예상금액',
-  price: 21700,
-  button: '주문하기(2개)',
-};
-
-export { ShoppingCartResult };
