--- src/component/index.js
@@ -1,7 +1,0 @@
-export { default as CartProduct } from 'component/CartProduct/CartProduct';
-export { default as ListProduct } from 'component/ListProduct/ListProduct';
-export { default as LoadingSpinner } from 'component/LoadingSpinner/LoadingSpinner';
-export { default as NavBar } from 'component/NavBar/NavBar';
-export { default as PageLayout } from 'component/PageLayout/PageLayout';
-export { default as Result } from 'component/Result/Result';
-export { default as Selector } from 'component/Selector/Selector';
