+++ src/components/@common/index.js
@@ -1,0 +1,4 @@
+export { default as Button } from 'components/@common/Button/Button';
+export { default as GridLayout } from 'components/@common/GridLayout/GridLayout';
+export { default as Header } from 'components/@common/Header/Header';
+export { default as Spinner } from 'components/@common/Spinner/Spinner';
