+++ src/components/LoadingSpinner/LoadingSpinner.stories.jsx
@@ -1,0 +1,15 @@
+import { LoadingSpinner } from 'components';
+
+export default {
+  title: 'Components/LoadingSpinner',
+  component: LoadingSpinner,
+  parameters: {
+    layout: 'centered',
+  },
+};
+
+const Template = args => <LoadingSpinner {...args} />;
+
+const DefaultLoadingSpinner = Template.bind({});
+
+export { DefaultLoadingSpinner };
