+++ src/components/index.js
@@ -1,0 +1,7 @@
+export { default as CartProduct } from 'components/CartProduct/CartProduct';
+export { default as ListProduct } from 'components/ListProduct/ListProduct';
+export { default as LoadingSpinner } from 'components/LoadingSpinner/LoadingSpinner';
+export { default as NavBar } from 'components/NavBar/NavBar';
+export { default as PageLayout } from 'components/PageLayout/PageLayout';
+export { default as Result } from 'components/Result/Result';
+export { default as Selector } from 'components/Selector/Selector';
