--- src/page/index.js
@@ -1,3 +1,0 @@
-export { default as NotFound } from 'page/NotFound/NotFound';
-export { default as ProductList } from 'page/ProductList/ProductList';
-export { default as ShoppingCart } from 'page/ShoppingCart/ShoppingCart';
