+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as NotFound } from 'pages/NotFound/NotFound';
+export { default as ProductList } from 'pages/ProductList/ProductList';
+export { default as ShoppingCart } from 'pages/ShoppingCart/ShoppingCart';
